Refactor rAthenaImport file structure
