--- dev/rAthenaImport/ParsedItem.ts
@@ -1,3 +1,0 @@
-import { Item } from "../../src/state/Item";
-
-export type ParsedItem = Item;
