+++ dev/rAthenaImport/generator/generateDrops.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import { dropsTemplate } from "./dropsTemplate";
-import { ParsedDrop } from "./ParsedDrop";
-import { MobId } from "../../src/state/MobId";
+import { ParsedDrop } from "../types/ParsedDrop";
+import { MobId } from "../../../src/state/MobId";
 
 export const generateDrops = (
   outputFile: string,
