+++ dev/rAthenaImport/index.ts
@@ -1,18 +1,19 @@
 import * as fs from "fs";
 import * as path from "path";
-import { concatLists } from "./concatLists";
-import { readJSONSync } from "./readJSONSync";
-import { getTableData } from "./getTableData";
-import { parseDataFile } from "./parseDataFile";
-import { parseItemsFromItemDB } from "./parseItemsFromItemDB";
-import { parseDropsFromMobDB } from "./parseDropsFromMobDB";
-import { parseDropsFromMobDrop } from "./parseDropsFromMobDrop";
-import { getMobIdsToImport } from "./getMobIdsToImport";
+import { concatLists } from "./util/concatLists";
+import { readJSONSync } from "./util/readJSONSync";
+import { parseTableData } from "./parser/parseTableData";
+import { parseDataFile } from "./parser/parseDataFile";
+import { parseItemsFromItemDB } from "./parser/parseItemsFromItemDB";
+import { parseDropsFromMobDB } from "./parser/parseDropsFromMobDB";
+import { parseDropsFromMobDrop } from "./parser/parseDropsFromMobDrop";
+import { getMobIdsToImport } from "./util/getMobIdsToImport";
 import { selectDrops } from "./selectDrops";
 import { selectItems } from "./selectItems";
-import { generateItems } from "./generateItems";
-import { generateDrops } from "./generateDrops";
+import { generateItems } from "./generator/generateItems";
+import { generateDrops } from "./generator/generateDrops";
 
+const mobsFile = path.resolve(__dirname, "../../src/fixtures/mobs.ts");
 const outputFolder = path.resolve(__dirname, "../../src/fixtures/generated");
 const inputFolder = process.argv[2];
 
@@ -25,7 +26,7 @@ const allItems = concatLists(
   [
     path.resolve(inputFolder, "item_db_re.json"),
     path.resolve(inputFolder, "item_db2_re.json"),
-  ].map((file) => parseItemsFromItemDB(getTableData(readJSONSync(file))))
+  ].map((file) => parseItemsFromItemDB(parseTableData(readJSONSync(file))))
 );
 
 const allDrops = parseDropsFromMobDB(
@@ -33,7 +34,7 @@ const allDrops = parseDropsFromMobDB(
     [
       path.resolve(inputFolder, "mob_db_re.json"),
       path.resolve(inputFolder, "mob_db2_re.json"),
-    ].map((file) => getTableData(readJSONSync(file)))
+    ].map((file) => parseTableData(readJSONSync(file)))
   )
 );
 
@@ -44,7 +45,7 @@ parseDropsFromMobDrop(
   )
 );
 
-const mobIds = getMobIdsToImport();
+const mobIds = getMobIdsToImport(mobsFile);
 const referencedDrops = selectDrops(allDrops, mobIds);
 const referencedItems = selectItems(allItems, referencedDrops);
