+++ dev/rAthenaImport/parser/parseDropsFromMobDB.ts
@@ -1,9 +1,9 @@
-import { RAthenaMob } from "./RAthenaMob";
-import { MobId } from "../../src/state/MobId";
-import { ParsedDrop } from "./ParsedDrop";
-import { ItemId } from "../../src/state/ItemId";
+import { RAthenaMob } from "../types/RAthenaMob";
+import { MobId } from "../../../src/state/MobId";
+import { ParsedDrop } from "../types/ParsedDrop";
+import { ItemId } from "../../../src/state/ItemId";
 import { parseChance } from "./parseChance";
-import { pull } from "./pull";
+import { pull } from "../util/pull";
 
 /**
  * Parses the data structure in rAthena database table mob_db
