+++ dev/rAthenaImport/parser/parseItemsFromItemDB.ts
@@ -1,6 +1,6 @@
-import { RAthenaItem } from "./RAthenaItem";
-import { ParsedItem } from "./ParsedItem";
-import { ItemId } from "../../src/state/ItemId";
+import { RAthenaItem } from "../types/RAthenaItem";
+import { ParsedItem } from "../types/ParsedItem";
+import { ItemId } from "../../../src/state/ItemId";
 
 /**
  * Parses the data structure in rAthena database table item_db
