+++ dev/rAthenaImport/types/ParsedItem.ts
@@ -1,0 +1,3 @@
+import { Item } from "../../../src/state/Item";
+
+export type ParsedItem = Item;
