+++ dev/rAthenaImport/util/readJSONSync.ts
